requirements.txt
aiofiles==0.7.0
anyio==3.3.2
asgiref==3.4.1
asn1crypto==1.4.0
atomicwrites==1.4.0
attrs==21.2.0
backports.zoneinfo==0.2.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
ci-info==0.2.0
click==8.0.1
colorama==0.4.4
configobj==5.0.6
configparser==5.0.2
cryptography==3.4.8
Django==3.2.7
django-crispy-forms==1.13.0
etelemetry==0.2.2
filelock==3.1.0
fpdf2==2.4.5
future==0.18.2
h11==0.12.0
httplib2==0.19.1
idna==3.2
iniconfig==1.1.1
isodate==0.6.0
itsdangerous==2.0.1
lxml==4.6.3
networkx==2.5.1
nibabel==3.2.1
nipype==1.6.1
numpy==1.19.5
opencv-python==4.5.3.56
oscrypto==1.2.1
packaging==21.0
pandas==1.1.5
pathlib==1.0.1
Pillow==8.3.2
pluggy==1.0.0
prov==2.0.0
psycopg2-binary==2.9.1
py==1.10.0
pycparser==2.20
pydot==1.4.2
pyHanko==0.8.0
pyhanko-certvalidator==0.16.0
PyMuPDF==1.16.14
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.2.5
pytest-django==4.4.0
python-dateutil==2.8.2
python-dotenv==0.19.0
pytils==0.3
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyxnat==1.4
PyYAML==6.0
qrcode==7.3.1
rdflib==5.0.0
reportlab==3.6.1
requests==2.26.0
scipy==1.5.4
simplejson==3.17.5
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.16.0
toml==0.10.2
traits==6.2.0
tzdata==2021.4
tzlocal==4.0.1
uritools==3.0.2
urllib3==1.26.7
uvicorn==0.15.0